task2_3_4_quality_xai/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
requests>=2.31.0
numpy>=1.26.0
Pillow>=10.0.0
pytest>=8.0.0
httpx>=0.27.0

tensorflow>=2.15.0
scikit-learn>=1.4.0

opencv-python-headless>=4.9.0
